+++ package.json
@@ -63,6 +63,7 @@
     "@supabase/auth-helpers-nextjs": "^0.10.0",
     "@supabase/supabase-js": "^2.45.4",
     "@t3-oss/env-nextjs": "^0.11.1",
+    "@tailwindcss/typography": "^0.5.16",
     "@tanstack/react-query": "^5.85.3",
     "@trpc/client": "^11.0.0-rc.553",
     "@trpc/next": "^11.0.0-rc.553",
@@ -74,6 +75,7 @@
     "clsx": "^2.1.1",
     "langchain": "^0.3.5",
     "lucide-react": "^0.539.0",
+    "marked": "^16.2.1",
     "next": "^15.4.6",
     "react": "19.0.0",
     "react-dom": "19.0.0",
